Move class FeatureComputerTestUtils to package org.mastodon.mamut.feature
